+++ rules/nette/tests/Rector/MethodCall/MergeDefaultsInGetConfigCompilerExtensionRector/Fixture/skip_empty_get_config.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Nette\Tests\Rector\MethodCall\GetConfigWithDefaultsArgumentToArrayMergeInCompilerExtensionRector\Fixture;
+namespace Rector\Nette\Tests\Rector\MethodCall\MergeDefaultsInGetConfigCompilerExtensionRector\Fixture;
 
 use Nette\DI\CompilerExtension;
